kubespand: restructure QEMU tests into separate targets
Split monolithic qemu_test.go into separate test directories with
independent Bazel go_test targets for parallel execution:
- nft/ (TestNftSmoke)
- kubespan/ (TestFlat, TestCrossSubnet, TestDiscoveryOnly)
- doublenat/ (TestDoubleNAT)
- talos/ (TestTalosKubeSpanDoubleNAT)

Each VM role gets its own init binary under vms/. Shared code in
initlib (PID-1 basics), kubespanlib (kubespand + probes), and the
qemu_tests package (VM management, events, test helpers).

Adds Talos diagnostic test that boots real Talos VMs from nocloud
qcow2 disk images to verify whether upstream KubeSpan can establish
WireGuard tunnels through double NAT.

See HANDOFF.md for status, gotchas, and next steps.
