+++ cluster/.gitignore
@@ -11,6 +11,8 @@ talos/home/generated/
 # Sensitive cluster credentials - generated by terraform
 kubeconfig*
 talosconfig*
+# Exception: test-only talosconfig with throwaway keys (not real cluster credentials)
+!kubespand/qemu_tests/talos/testdata/vps/talosconfig
 
 # Nix cache signing key - generated by nix-store, cached locally
 terraform/bootstrap/persistent-auth/nix-cache-key.json
