--- cluster/kubespand/qemu/BUILD.bazel
@@ -1,165 +1,0 @@
-# QEMU-based nftables and KubeSpan integration tests.
-#
-# Test functions in a single go_test:
-#   TestNftSmoke:             nft-smoke levels 1-6 in a single QEMU VM
-#   TestKubeSpanFlat:         flat topology (same subnet, IPv6 ULA probe)
-#   TestKubeSpanCrossSubnet:  cross-subnet topology (rpfilter + IPv4 probe)
-#   TestKubeSpanDiscoveryOnly: discovery-only topology (no extra_endpoints)
-#   TestKubeSpanDoubleNAT:    double-NAT topology (5 VMs, realistic NAT traversal)
-#
-# Two initramfs images:
-#   base:      Alpine minirootfs + init + kubespand + kernel modules
-#   discovery: Alpine minirootfs + init + discovery-service + kernel modules (no kubespand)
-# No Docker dependency — discovery runs in a QEMU VM.
-# QEMU TCG (software emulation) is used.
-
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-# ─ Shared event schema ────────────────────────────────────────────────────
-# Imported by both qemu_init (VM-side, emits events) and qemu_test (host-side, parses events).
-go_library(
-    name = "qemu",
-    testonly = True,
-    srcs = ["events.go"],
-    importpath = "github.com/agentydragon/ducktape/cluster/kubespand/qemu",
-    visibility = ["//cluster/kubespand:__subpackages__"],
-)
-
-# ─ Extract vmlinuz from Alpine linux-virt .apk ────────────────────────────────
-genrule(
-    name = "vmlinuz",
-    srcs = ["@alpine_linux_virt//file"],
-    outs = ["vmlinuz-virt"],
-    cmd = "tar -xzf $< -O boot/vmlinuz-virt > $@",
-)
-
-# ─ Extract full kernel modules tree for modprobe ─────────────────────────────
-# Extracts the entire /lib/modules/<version>/ tree with modules.dep so busybox
-# modprobe can resolve dependencies automatically. Used by both nft-smoke
-# (nftables modules) and kubespan (wireguard, nftables, virtio_net).
-genrule(
-    name = "modules_tree",
-    srcs = ["@alpine_linux_virt//file"],
-    outs = ["modules_tree.tar"],
-    cmd = """
-        set -e
-        TMPDIR=$$(mktemp -d)
-        # Extract kernel modules from Alpine .apk.
-        tar -xzf $< -C $$TMPDIR lib/modules/
-        # Decompress all .ko.gz files for use with insmod/modprobe.
-        find $$TMPDIR/lib/modules/ -name '*.ko.gz' -exec gunzip {} \\;
-        # Update modules.dep and modules.dep.bin to reference .ko instead of .ko.gz.
-        # busybox modprobe reads modules.dep to resolve dependencies.
-        find $$TMPDIR/lib/modules/ -name 'modules.dep' -exec sed -i 's/\\.ko\\.gz/.ko/g' {} \\;
-        find $$TMPDIR/lib/modules/ -name 'modules.alias' -exec sed -i 's/\\.ko\\.gz/.ko/g' {} \\;
-        # Regenerate binary index files from the text files.
-        # modules.dep.bin is a binary format; remove it so modprobe falls back to text.
-        find $$TMPDIR/lib/modules/ -name '*.bin' -delete
-        # Create a tarball of the modules tree (preserving directory structure).
-        tar -cf $@ -C $$TMPDIR lib/modules/
-        rm -rf $$TMPDIR
-    """,
-)
-
-# ─ Build initramfs (cpio.gz) ────────────────────────────────────────────────
-# Base initramfs for kubespan, router, and nft_smoke VMs. Uses Alpine minirootfs
-# as the base filesystem (provides /bin, /sbin, /etc, busybox + symlinks, musl libc).
-# Overlays: /init (Go binary), /kubespand, kernel modules.
-# The init binary dispatches on mode= kernel cmdline parameter.
-genrule(
-    name = "initramfs",
-    testonly = True,
-    srcs = [
-        "@alpine_minirootfs_x86_64//file",
-        "//cluster/kubespand",
-        "//cluster/kubespand/qemu/qemu_init",
-        ":modules_tree",
-    ],
-    outs = ["initramfs.cpio.gz"],
-    cmd = """
-        set -e
-        INITRAMFS_DIR=$$(mktemp -d)
-
-        # Extract Alpine minirootfs as base (provides filesystem layout + busybox symlinks).
-        tar -xzf $(location @alpine_minirootfs_x86_64//file) -C $$INITRAMFS_DIR
-
-        # Create directories not in minirootfs.
-        mkdir -p $$INITRAMFS_DIR/var/lib/kubespan \
-                 $$INITRAMFS_DIR/etc/kubespan
-
-        # Install init binary (Go, replaces init.sh).
-        cp $(location //cluster/kubespand/qemu/qemu_init) $$INITRAMFS_DIR/init
-        chmod +x $$INITRAMFS_DIR/init
-
-        # Install kubespand.
-        cp $(location //cluster/kubespand:kubespand) $$INITRAMFS_DIR/kubespand
-        chmod +x $$INITRAMFS_DIR/kubespand
-
-        # Install full kernel modules tree (for modprobe dependency resolution).
-        tar -xf $(location :modules_tree) -C $$INITRAMFS_DIR
-
-        # Build cpio archive.
-        (cd $$INITRAMFS_DIR && find . -print0 | cpio --null -o -H newc 2>/dev/null) | gzip -9 > $@
-        rm -rf $$INITRAMFS_DIR
-    """,
-)
-
-# ─ Build discovery initramfs (cpio.gz) ──────────────────────────────────────
-# Minimal initramfs for discovery-only VMs. Includes the discovery-service
-# binary (~14MB) but NOT kubespand (discovery VMs don't need it). Kept
-# separate to avoid bloating the base initramfs beyond the 512MB QEMU RAM.
-genrule(
-    name = "initramfs_discovery",
-    testonly = True,
-    srcs = [
-        "@alpine_minirootfs_x86_64//file",
-        "@discovery_service_binary//file",
-        "//cluster/kubespand/qemu/qemu_init",
-        ":modules_tree",
-    ],
-    outs = ["initramfs-discovery.cpio.gz"],
-    cmd = """
-        set -e
-        INITRAMFS_DIR=$$(mktemp -d)
-
-        # Extract Alpine minirootfs as base (provides filesystem layout + busybox symlinks).
-        tar -xzf $(location @alpine_minirootfs_x86_64//file) -C $$INITRAMFS_DIR
-
-        # Install init binary (Go).
-        cp $(location //cluster/kubespand/qemu/qemu_init) $$INITRAMFS_DIR/init
-        chmod +x $$INITRAMFS_DIR/init
-
-        # Install discovery service.
-        cp $(location @discovery_service_binary//file) $$INITRAMFS_DIR/discovery-service
-        chmod +x $$INITRAMFS_DIR/discovery-service
-
-        # Install kernel modules (virtio_net needed for networking).
-        tar -xf $(location :modules_tree) -C $$INITRAMFS_DIR
-
-        # Build cpio archive.
-        (cd $$INITRAMFS_DIR && find . -print0 | cpio --null -o -H newc 2>/dev/null) | gzip -9 > $@
-        rm -rf $$INITRAMFS_DIR
-    """,
-)
-
-# ─ Test: QEMU integration tests ────────────────────────────────────────────
-go_test(
-    name = "qemu_test",
-    timeout = "eternal",
-    srcs = ["qemu_test.go"],
-    data = [
-        ":initramfs",
-        ":initramfs_discovery",
-        ":vmlinuz",
-        "@talos_initramfs_amd64//file",
-        "@talos_vmlinuz_amd64//file",
-        "@talosctl_amd64//file",
-    ],
-    embed = [":qemu"],
-    tags = [
-        "requires_qemu",
-    ],
-    deps = [
-        "@io_bazel_rules_go//go/runfiles",
-    ],
-)
