--- cluster/kubespand/qemu/qemu_init/BUILD.bazel
@@ -1,40 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
-
-go_library(
-    name = "qemu_init_lib",
-    testonly = True,
-    srcs = [
-        "main.go",
-        "mode_discovery.go",
-        "mode_kubespan.go",
-        "mode_nft_smoke.go",
-        "mode_router.go",
-        "nft_exprs.go",
-        "nft_smoke.go",
-        "ping.go",
-    ],
-    importpath = "github.com/agentydragon/ducktape/cluster/kubespand/qemu/qemu_init",
-    visibility = ["//visibility:private"],
-    deps = [
-        "//cluster/kubespand/agentconfig",
-        "//cluster/kubespand/qemu",
-        "@com_github_google_nftables//:nftables",
-        "@com_github_google_nftables//binaryutil",
-        "@com_github_google_nftables//expr",
-        "@in_gopkg_yaml_v3//:yaml_v3",
-        "@org_golang_x_net//icmp",
-        "@org_golang_x_net//ipv4",
-        "@org_golang_x_net//ipv6",
-        "@org_golang_x_sys//unix",
-    ],
-)
-
-go_binary(
-    name = "qemu_init",
-    testonly = True,
-    embed = [":qemu_init_lib"],
-    goarch = "amd64",
-    goos = "linux",
-    static = "on",
-    visibility = ["//cluster/kubespand:__subpackages__"],
-)
