+++ cluster/kubespand/qemu_tests/talos/testdata/vps/talosconfig
@@ -1,0 +1,7 @@
+context: test-kubespan
+contexts:
+    test-kubespan:
+        endpoints: []
+        ca: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUJQakNCOGFBREFnRUNBaEJoa3NtcjZIY3dBUG94cWE3aWpQcmVNQVVHQXl0bGNEQVFNUTR3REFZRFZRUUsKRXdWMFlXeHZjekFlRncweU5qQXpNVFF3T1RVeE1EVmFGdzB6TmpBek1URXdPVFV4TURWYU1CQXhEakFNQmdOVgpCQW9UQlhSaGJHOXpNQ293QlFZREsyVndBeUVBamdWbldRTzd1TkZWdzN2MktQdHVDK0FncGpPRmdjZUZVeEVwCndUMmdTSVNqWVRCZk1BNEdBMVVkRHdFQi93UUVBd0lDaERBZEJnTlZIU1VFRmpBVUJnZ3JCZ0VGQlFjREFRWUkKS3dZQkJRVUhBd0l3RHdZRFZSMFRBUUgvQkFVd0F3RUIvekFkQmdOVkhRNEVGZ1FVZWw4OVRUaHJRUlVYY0h3aQpUdUo4ZUdqbXZTTXdCUVlESzJWd0EwRUFMcUxlL3lIQ1pNRmpiL0hmQ1BVUDRvTnZVc1BBS3EwTFg5RGhKUkVrCjdwOXFIbGE1TVBwb3piODUrM09HUzE1clR6dmRrdGNKY3Y4L0tFdGpsNG5kQ1E9PQotLS0tLUVORCBDRVJUSUZJQ0FURS0tLS0tCg==
+        crt: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUJLRENCMjZBREFnRUNBaEJKMlNDREdkeGZkZFg2L0M3RzB5VTJNQVVHQXl0bGNEQVFNUTR3REFZRFZRUUsKRXdWMFlXeHZjekFlRncweU5qQXpNVFF3T1RVeE1EWmFGdzB5TnpBek1UUXdPVFV4TURaYU1CTXhFVEFQQmdOVgpCQW9UQ0c5ek9tRmtiV2x1TUNvd0JRWURLMlZ3QXlFQUQ5bDVOSHpqRGlDYm1vTFJjd2l3aUM2ZllBRjdJdjcwCnZJL3lEbVZINVl5alNEQkdNQTRHQTFVZER3RUIvd1FFQXdJSGdEQVRCZ05WSFNVRUREQUtCZ2dyQmdFRkJRY0QKQWpBZkJnTlZIU01FR0RBV2dCUjZYejFOT0d0QkZSZHdmQ0pPNG54NGFPYTlJekFGQmdNclpYQURRUUIxa29QWQpENE0wNVJUZmNXMURuU21NNEFqcVFqbU9IV09OU3RoV0lDSkNzVkhYNnFheEswSjZUTEYwcWo3MjBZdWVDRTdOCi9jZy9WUDVwQ243M1hvVUkKLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=
+        key: LS0tLS1CRUdJTiBFRDI1NTE5IFBSSVZBVEUgS0VZLS0tLS0KTUM0Q0FRQXdCUVlESzJWd0JDSUVJSVN4WTZ5OUwzMVpKYWxoV0lJU2F2RXVDUTJQT0t4bVBwQ0E1TlRKKzBjMwotLS0tLUVORCBFRDI1NTE5IFBSSVZBVEUgS0VZLS0tLS0K
