+++ cluster/kubespand/qemu_tests/vms/initlib/BUILD.bazel
@@ -1,0 +1,10 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "initlib",
+    testonly = True,
+    srcs = ["initlib.go"],
+    importpath = "github.com/agentydragon/ducktape/cluster/kubespand/qemu_tests/vms/initlib",
+    visibility = ["//cluster/kubespand:__subpackages__"],
+    deps = ["//cluster/kubespand/qemu_tests"],
+)
