+++ cluster/kubespand/qemu_tests/vms/kubespanlib/BUILD.bazel
@@ -1,0 +1,21 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "kubespanlib",
+    testonly = True,
+    srcs = [
+        "kubespanlib.go",
+        "probes.go",
+    ],
+    importpath = "github.com/agentydragon/ducktape/cluster/kubespand/qemu_tests/vms/kubespanlib",
+    visibility = ["//cluster/kubespand:__subpackages__"],
+    deps = [
+        "//cluster/kubespand/agentconfig",
+        "//cluster/kubespand/qemu_tests",
+        "//cluster/kubespand/qemu_tests/vms/initlib",
+        "@in_gopkg_yaml_v3//:yaml_v3",
+        "@org_golang_x_net//icmp",
+        "@org_golang_x_net//ipv4",
+        "@org_golang_x_net//ipv6",
+    ],
+)
