+++ cluster/kubespand/qemu_tests/vms/kubespanlib/probes.go
@@ -1,4 +1,4 @@
-package main
+package kubespanlib
 
 import (
 	"fmt"
@@ -9,11 +9,13 @@ import (
 	"golang.org/x/net/icmp"
 	"golang.org/x/net/ipv4"
 	"golang.org/x/net/ipv6"
+
+	qemu_tests "github.com/agentydragon/ducktape/cluster/kubespand/qemu_tests"
+	"github.com/agentydragon/ducktape/cluster/kubespand/qemu_tests/vms/initlib"
 )
 
-// tcpProbe attempts a TCP connection to target:port with retry until timeout.
-// Returns true on first successful connection.
-func tcpProbe(target string, port int, timeout time.Duration) bool {
+// TCPProbe attempts a TCP connection to target:port with retry until timeout.
+func TCPProbe(target string, port int, timeout time.Duration) bool {
 	addr := net.JoinHostPort(target, fmt.Sprintf("%d", port))
 	deadline := time.Now().Add(timeout)
 
@@ -31,13 +33,8 @@ func tcpProbe(target string, port int, timeout time.Duration) bool {
 	return false
 }
 
-// serveTCP starts a TCP listener on the given port that accepts connections
-// and immediately closes them (for probe testing). Runs until the returned
-// cancel function is called.
-// serveTCP starts TCP listeners on the given port on both IPv4 and IPv6.
-// Accepts connections and immediately closes them (for probe testing).
-// Runs until the returned cancel function is called.
-func serveTCP(port int) (cancel func()) {
+// ServeTCP starts TCP listeners on the given port on both IPv4 and IPv6.
+func ServeTCP(port int) (cancel func()) {
 	addr := fmt.Sprintf(":%d", port)
 	var listeners []net.Listener
 	for _, network := range []string{"tcp4", "tcp6"} {
@@ -64,10 +61,8 @@ func serveTCP(port int) (cancel func()) {
 	}
 }
 
-// probe sends ICMP echo requests to the target address with retry until
-// timeout. Auto-detects IPv4 vs IPv6 from the target. Returns true on
-// first successful echo reply.
-func probe(target string, timeout time.Duration) bool {
+// Probe sends ICMP echo requests to the target with retry until timeout.
+func Probe(target string, timeout time.Duration) bool {
 	deadline := time.Now().Add(timeout)
 	seq := 0
 
@@ -93,7 +88,6 @@ func probe(target string, timeout time.Duration) bool {
 	return false
 }
 
-// ping4 sends a single ICMPv4 echo request and waits up to 3s for a reply.
 func ping4(target string, seq int) bool {
 	conn, err := icmp.ListenPacket("ip4:icmp", "0.0.0.0")
 	if err != nil {
@@ -107,13 +101,8 @@ func ping4(target string, seq int) bool {
 	}
 
 	msg := icmp.Message{
-		Type: ipv4.ICMPTypeEcho,
-		Code: 0,
-		Body: &icmp.Echo{
-			ID:   os.Getpid() & 0xffff,
-			Seq:  seq,
-			Data: []byte("kubespan-probe"),
-		},
+		Type: ipv4.ICMPTypeEcho, Code: 0,
+		Body: &icmp.Echo{ID: os.Getpid() & 0xffff, Seq: seq, Data: []byte("kubespan-probe")},
 	}
 	wb, err := msg.Marshal(nil)
 	if err != nil {
@@ -122,10 +111,8 @@ func ping4(target string, seq int) bool {
 
 	dst, err := net.ResolveIPAddr("ip4", target)
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "resolve %s: %v\n", target, err)
 		return false
 	}
-
 	if _, err := conn.WriteTo(wb, dst); err != nil {
 		return false
 	}
@@ -135,16 +122,13 @@ func ping4(target string, seq int) bool {
 	if err != nil {
 		return false
 	}
-
-	rm, err := icmp.ParseMessage(1, rb[:n]) // 1 = ICMPv4 protocol number
+	rm, err := icmp.ParseMessage(1, rb[:n])
 	if err != nil {
 		return false
 	}
-
 	return rm.Type == ipv4.ICMPTypeEchoReply
 }
 
-// ping6 sends a single ICMPv6 echo request and waits up to 3s for a reply.
 func ping6(target string, seq int) bool {
 	conn, err := icmp.ListenPacket("ip6:ipv6-icmp", "::")
 	if err != nil {
@@ -158,13 +142,8 @@ func ping6(target string, seq int) bool {
 	}
 
 	msg := icmp.Message{
-		Type: ipv6.ICMPTypeEchoRequest,
-		Code: 0,
-		Body: &icmp.Echo{
-			ID:   os.Getpid() & 0xffff,
-			Seq:  seq,
-			Data: []byte("kubespan-probe"),
-		},
+		Type: ipv6.ICMPTypeEchoRequest, Code: 0,
+		Body: &icmp.Echo{ID: os.Getpid() & 0xffff, Seq: seq, Data: []byte("kubespan-probe")},
 	}
 	wb, err := msg.Marshal(nil)
 	if err != nil {
@@ -173,10 +152,8 @@ func ping6(target string, seq int) bool {
 
 	dst, err := net.ResolveIPAddr("ip6", target)
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "resolve %s: %v\n", target, err)
 		return false
 	}
-
 	if _, err := conn.WriteTo(wb, dst); err != nil {
 		return false
 	}
@@ -186,11 +163,39 @@ func ping6(target string, seq int) bool {
 	if err != nil {
 		return false
 	}
-
-	rm, err := icmp.ParseMessage(58, rb[:n]) // 58 = ICMPv6 protocol number
+	rm, err := icmp.ParseMessage(58, rb[:n])
 	if err != nil {
 		return false
 	}
-
 	return rm.Type == ipv6.ICMPTypeEchoReply
+}
+
+// EmitProbe emits a probe event and dumps kubespand log on failure.
+func EmitProbe(msg, target string, ok bool) {
+	evt := qemu_tests.Event{Type: qemu_tests.EventProbe, Message: msg, Target: target, Success: &ok}
+	if !ok {
+		evt.Error = "probe failed"
+		initlib.DumpLog("/tmp/kubespand.log")
+	}
+	initlib.EmitEvent(evt)
+}
+
+// RunProbes runs the standard 2-node probe suite (IPv6 ULA + IPv4 bridge, ICMP + TCP).
+func RunProbes(peerAddr, peerBridgeIP string, tcpPort int) {
+	EmitProbe("ipv6 ULA icmp", peerAddr, Probe(peerAddr, 60*time.Second))
+	EmitProbe("ipv4 peer eth0 icmp", peerBridgeIP, Probe(peerBridgeIP, 60*time.Second))
+	EmitProbe("ipv6 ULA tcp", fmt.Sprintf("[%s]:%d", peerAddr, tcpPort),
+		TCPProbe(peerAddr, tcpPort, 30*time.Second))
+	EmitProbe("ipv4 peer eth0 tcp", fmt.Sprintf("%s:%d", peerBridgeIP, tcpPort),
+		TCPProbe(peerBridgeIP, tcpPort, 30*time.Second))
+}
+
+// RunDoubleNATProbes probes each peer's ULA via ICMP and TCP.
+func RunDoubleNATProbes(peerAddrs []string, tcpPort int) {
+	for i, addr := range peerAddrs {
+		label := fmt.Sprintf("peer %d", i+1)
+		EmitProbe(label+" ULA icmp", addr, Probe(addr, 60*time.Second))
+		EmitProbe(label+" ULA tcp", fmt.Sprintf("[%s]:%d", addr, tcpPort),
+			TCPProbe(addr, tcpPort, 30*time.Second))
+	}
 }
