+++ cluster/kubespand/qemu_tests/vms/router/BUILD.bazel
@@ -1,0 +1,55 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
+
+go_library(
+    name = "router_lib",
+    testonly = True,
+    srcs = ["main.go"],
+    importpath = "github.com/agentydragon/ducktape/cluster/kubespand/qemu_tests/vms/router",
+    visibility = ["//visibility:private"],
+    deps = [
+        "//cluster/kubespand/qemu_tests",
+        "//cluster/kubespand/qemu_tests/vms/initlib",
+        "@com_github_google_nftables//:nftables",
+        "@com_github_google_nftables//expr",
+    ],
+)
+
+go_binary(
+    name = "router",
+    testonly = True,
+    embed = [":router_lib"],
+    goarch = "amd64",
+    goos = "linux",
+    static = "on",
+    visibility = ["//cluster/kubespand:__subpackages__"],
+)
+
+genrule(
+    name = "initramfs",
+    testonly = True,
+    srcs = [
+        "@alpine_minirootfs_x86_64//file",
+        ":router",
+        "//cluster/kubespand/qemu_tests:modules_tree",
+    ],
+    outs = ["initramfs.cpio.gz"],
+    cmd = """
+        set -e
+        INITRAMFS_DIR=$$(mktemp -d)
+
+        # Extract Alpine minirootfs as base (provides filesystem layout + busybox symlinks).
+        tar -xzf $(location @alpine_minirootfs_x86_64//file) -C $$INITRAMFS_DIR
+
+        # Install init binary (Go).
+        cp $(location :router) $$INITRAMFS_DIR/init
+        chmod +x $$INITRAMFS_DIR/init
+
+        # Install kernel modules (nftables, virtio_net needed).
+        tar -xf $(location //cluster/kubespand/qemu_tests:modules_tree) -C $$INITRAMFS_DIR
+
+        # Build cpio archive.
+        (cd $$INITRAMFS_DIR && find . -print0 | cpio --null -o -H newc 2>/dev/null) | gzip -9 > $@
+        rm -rf $$INITRAMFS_DIR
+    """,
+    visibility = ["//cluster/kubespand:__subpackages__"],
+)
